# h8ocbc2_milestone1_012/swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code
  version: "1.0.0"
  title: Swagger Rest Article
consumes:
  - application/json
produces:
  - application/json
basePath: /api
...
paths:
  /directors:
    get:
      operationId: directors.read_all
      tags:
        - Directors
      summary: Read the entire set of directors, sorted by id
      description: Read the entire set of directors, sorted by id
      parameters:
        - name: limit
          in: query
          type: integer
          description: number of directors displayed
      responses:
        200:
          description: Successfully read directors set operation
          schema:
            type: array
            items:
              properties:
                name:
                  type: string
                  description: Name of the director
                id:
                  type: integer
                  description: Id of the director
                gender:
                  type: integer
                  description: Gender of the director
                uid:
                  type: integer
                  description: UID of the director
                department:
                  type: string
                  description: Department of the director
                movies:
                  type: array
                  items:
                    properties:
                      director_id:
                        type: integer
                        description: Id of director this movie is associated with
                      id:
                        type: integer
                        description: Id of this movie
                      original_title:
                        type: string
                        description: Original title of this movie
                      budget:
                        type: integer
                        description: Budget of this movie
                      popularity:
                        type: integer
                        description: Popularity of this movie
                      release_date:
                        type: string
                        description: Release date of this movie
                      revenue:
                        type: integer
                        description: Revenue of this movie
                      title:
                        type: string
                        description: Title of this movie
                      vote_average:
                        type: number
                        description: Average vote of this movie
                      vote_count:
                        type: integer
                        description: Number of vote of this movie
                      overview:
                        type: string
                        description: Overview of this movie
                      tagline:
                        type: string
                        description: Tagline of this movie
                      uid:
                        type: integer
                        description: Uid of this movie
    
    post:
      operationId: directors.create
      tags:
        - Directors
      summary: Create a director
      description: Create a new director
      parameters:
        - name: director
          in: body
          description: Director to create
          required: True
          schema:
            type: object
            properties: 
              name: 
                type: string
                description: director name
              gender:
                type: integer
                description: director gender
              uid:
                type: integer
                description: director unique id
              department:
                type: string
                description: director department
      responses:
        201:
          description: Successfully created director
          schema:
            properties:
              id:
                type: integer
                description: director Id
              name: 
                type: string
                description: director name
              gender:
                type: integer
                description: director gender
              uid:
                type: integer
                description: director unique id
              department:
                type: string
                description: director department
...